requirements.txt
requests
beautifulsoup4
lxml
google-api-python-client
google-auth-oauthlib